+++ src/app/app-routing.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from '@app/home/home.component';
+import { SelectivePreloadStrategyService } from '@app/route-utility/selective-preload-strategy.service';
+import { AuthenticationGuard } from '@auth/guards/authentication.guard';
 import { AppShellComponent } from '@layout/app-shell/app-shell.component';
 
 const routes: Routes = [
@@ -8,9 +10,15 @@ const routes: Routes = [
     path: '',
     component: AppShellComponent,
     children: [
-      { path: 'home', component: HomeComponent },
+      {
+        path: 'home',
+        component: HomeComponent,
+        // TODO: Get proper page name & keys from SEO specialist and update pageTitle
+        data: { pageTitle: 'My Home' }
+      },
       {
         path: 'auth',
+        canLoad: [AuthenticationGuard],
         loadChildren: () =>
           import('@app/auth/my-profile/my-profile.module').then(
             m => m.MyProfileModule
@@ -18,21 +26,25 @@ const routes: Routes = [
       },
       {
         path: 'cart',
+        data: { preload: true, delay: true },
         loadChildren: () =>
           import('@app/cart/cart.module').then(m => m.CartModule)
       },
       {
         path: 'checkout',
+        canLoad: [AuthenticationGuard],
         loadChildren: () =>
           import('@app/checkout/checkout.module').then(m => m.CheckoutModule)
       },
       {
         path: 'payment',
+        canLoad: [AuthenticationGuard],
         loadChildren: () =>
           import('@app/payment/payment.module').then(m => m.PaymentModule)
       },
       {
         path: 'confirmation',
+        canLoad: [AuthenticationGuard],
         loadChildren: () =>
           import('@app/confirmation/confirmation.module').then(
             m => m.ConfirmationModule
@@ -60,6 +72,8 @@ const routes: Routes = [
 @NgModule({
   imports: [
     RouterModule.forRoot(routes, {
+      enableTracing: true, // <-- debugging purposes only
+      preloadingStrategy: SelectivePreloadStrategyService,
       scrollPositionRestoration: 'enabled',
       anchorScrolling: 'enabled'
     })