+++ src/app/checkout/checkout-routing.module.ts
@@ -1,11 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CheckoutComponent } from '@app/checkout/checkout.component';
+import { AuthenticationGuard } from '@auth/guards/authentication.guard';
 
 const routes: Routes = [
   {
     path: '',
-    component: CheckoutComponent
+    canActivate: [AuthenticationGuard],
+    component: CheckoutComponent,
+    data: { pageTitle: 'My Checkout' }
   }
 ];