+++ src/environments/environment.prod.ts
@@ -2,7 +2,7 @@ export const environment = {
   production: true,
   webAPIUrl: 'https://mydomain.com/api/v1',
   // version is automatically updated during build from package.json
-  version: '0.0.11',
+  version: '0.0.12',
   configFilePath: 'assets/config/prod.config.json',
   envName: 'prod',
   firestore: {